bash script to automate compilation and run(s)
This script helps setup and run. The first run of
$ ./workon 1
will create a folder '1', copy 'template.cpp' to that folder,
create an 'input' file and start watching for any changes
to any files on that folder. The subsequent runs will directly
jump to watching the files for changes.

If main.cpp changes, it is recompiled and run. If any other
file changes, main.cpp is run with the changed file as input.