CD - Subset Enumeration (Recursive Backtracking)
For the given tracks, the solution generates all subsets (with
some pruning if the current subset under generation already violates
necessary conditions) and checks if the current subset is the best
solution it has seen so far.

The recommended approach in the book is recursive backtracking. Might
attempt writing a recursive algorithm just for the sake of exercise.