00624 CD - Recursive Backtracking
Rewrote the algorithm to use recursion over subset enumeration to
as an exercise on recursive backtracking as suggested by the book.

The solution uses a solve function:
@params
remaining: the space remaining in the tape.
idx: the index and forward to look into songs durations from input
@return
least possible amount of space remaining in the tape
(as a side consequence, the global ans is set to the best solution)