10261 Ferry Loading -- DP (Top Down)
solved pretty much the standard knapsack way. The only hurdle was in
realizing that (id, lspace) along with total were sufficient parameters
to determine the state uniquely. This is because all the cars have to be
loaded sequentially and if one arrives at the (id, lspace) everything
else must have gone to rspace.

There were some additional corner cases to be considered while priting
solution (e.g. the case where there were zero cars on the queue and
the case when one has to know when all the possible cars that could be
loaded are loaded already by looking at the states).